requirements.txt
flask
psycopg2-binary
flask-sqlalchemy
flask-restful
